Add validations to the Cite extension
The class Validator has been redesigned. It has now five
methods, all called in Cite.php, corresponding to five
categories of errors:

1- The method validateNewRef() can be called as soon
as the attributes of the tag are processed, but only
when the tag is not in list. It is called in
gardedRef().

2- The method  validateNewRefInList() can also be
called as soon as the attributes of the tag are
processed, but when the tag is in list.  It is called
in gardedRef().

3- The method validateNewHalfParsedHtml() can be
called as soon as the content of the tag has been
processed. Also callled in gardedRef().

4- The method validateGroupReferences() can be called
as soon as all the ref items of a references tag has
been processed. It is called in references().

5- The method validateRemainingRef() is called at
the end after the page has been parsed. It is
called in checkRefsNoReferences(), which is called
in the onParserAfterParse() hook.

The code could be improved by having the methods
cumulate the warnings instead of returning as soon
as an error is detected. The tool used allows that.

Also, there is no claim that these validation
methods are exhaustive.