Remove fullyParsed() and rename variable in halfParsed()
The method fullyParsed() should not be needed, because
the extension must return half parsed html to the parser.
The method halfParsed return half parsed html. To make
the code easier to read, I renamed the variable for
the returned value in half parsed html (before wrapping
into a div) from $wikitext to $halfParsed.